refactor: extract download.js and history-recal.js from index.js
- download.js: setupDownloadForm() — wires the download form
- history-recal.js: addHistoryRecal() — history-mode recalculate button,
  uses direct state import instead of window.GTCGlobal
- index.js: 625 → 552 lines (73 removed)
- Remove now-unused downloadOutputFiles import from index.js
